support analyzing all refs

+++ js/iobio.js
@@ -16,29 +16,29 @@ var shortid = require('shortid');
 
 
 // Command function starts here
-iobio.cmd = function(service, params, opts) {	
+iobio.cmd = function(service, params, opts) {
 	// Call EventEmitter constructor
 	EventEmitter.call(this);
-	
+
 	var extend = require('extend');
-	
+
    	this.options = {
-   		/* defaults */ 
+   		/* defaults */
    		id: shortid.generate()
    	};
-   	extend(this.options, opts);      	
-	this.protocol = 'ws';	
-	this.pipedCommands = { };	
+   	extend(this.options, opts);
+	this.protocol = this.options.ssl ? 'wss' : 'ws';
+	this.pipedCommands = { };
 	this.pipedCommands[ this.options.id ] = this;
 
 	// make sure params isn't undefined
 	params = params || [];
 
-	var conn = require('./conn.js'); // handles connection code		
+	var conn = require('./conn.js'); // handles connection code
 	this.connection = new conn(this.protocol, service, params, this.options);
 	var me = this;
-	
-	// bind stream events	
+
+	// bind stream events
 	require('./utils/bindStreamEvents')(this, this.connection);
 }
 
@@ -48,10 +48,10 @@ inherits(iobio.cmd, EventEmitter);
 // functions
 
 // Chain commands
-iobio.cmd.prototype.pipe = function(service, params, opts) {	
-	
-	
-	// add current url to params		
+iobio.cmd.prototype.pipe = function(service, params, opts) {
+
+
+	// add current url to params
 	params = params || [];
 	params.push( this.url() );
 
@@ -61,15 +61,17 @@ iobio.cmd.prototype.pipe = function(service, params, opts) {
 
 	// transfer pipedCommands to new command;
 	for (var id in this.pipedCommands ) { newCmd.pipedCommands[id] = this.pipedCommands[id]; }
-	
+
 	return newCmd;
 }
 
 // Create url
 iobio.cmd.prototype.url = function() { return 'iobio://' + this.connection.uri; }
 iobio.cmd.prototype.http = function() { return 'http://' + this.connection.uri; }
+iobio.cmd.prototype.https = function() { return 'https://' + this.connection.uri; }
 iobio.cmd.prototype.ws = function() { return 'ws://' + this.connection.uri; }
-iobio.cmd.prototype.id = this.id 
+iobio.cmd.prototype.wss = function() { return 'wss://' + this.connection.uri; }
+iobio.cmd.prototype.id = this.id
 
 
 // getters/setters
@@ -80,23 +82,31 @@ iobio.cmd.prototype.protocol = function(_) {
 }
 
 // Execute command
-iobio.cmd.prototype.run = function() {	
+iobio.cmd.prototype.run = function() {
 
- 	// run 
-	this.connection.run(this.pipedCommands); 
+ 	// run
+	this.connection.run(this.pipedCommands);
 	// send pipedCommands so that each command can properly handle the request comming back
-	// e.g. if the second command of 3 that are piped together is sending file data then it 
+	// e.g. if the second command of 3 that are piped together is sending file data then it
 	// should handle the request for data coming from the server.
+}
+
+// Close client and let server handle cleanup
+iobio.cmd.prototype.closeClient = function() {
+	if (this.connection && this.connection.runner )
+		this.connection.runner.closeClient();
 }
 
 // Kill running command instantly, leaving any data in the pipe
 iobio.cmd.prototype.kill = function() {
-	this.connection.runner.kill();
+	if (this.connection && this.connection.runner )
+		this.connection.runner.kill();
 }
 
 // End command safely. This may take a second or two and still give more data
 iobio.cmd.prototype.end = function() {
-	this.connection.runner.end();
+	if (this.connection && this.connection.runner )
+		this.connection.runner.end();
 }
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
 
@@ -4136,26 +4146,26 @@ module.exports = 0;
 var EventEmitter = require('events').EventEmitter;
 var inherits = require('inherits');
 
-var conn = function(protocol, service, params, opts) {	
+var conn = function(protocol, service, params, opts) {
 	// Call EventEmitter constructor
 	EventEmitter.call(this);
 
-	var me = this;	
+	var me = this;
 	this.opts = opts;
 	this.service = service;
 	this.protocol = protocol;
-	this.params = params;	
+	this.params = params;
 
-	// create url	
+	// create url
 	var UrlBuilder = require('./urlBuilder.js');
 	var	urlBuilder = new UrlBuilder(service, params, opts);
 	this.urlBuilder = urlBuilder;
 	this.uri = urlBuilder.uri;
 
-	if (protocol == 'ws')
+	if (protocol == 'ws' || protocol == 'wss')
 		this.Runner  = require('./protocol/ws.js');
-	else if (protocol == 'http')
-		this.Runner = require('./protocol/http.js');	
+	else if (protocol == 'http' || protocol == 'https')
+		this.Runner = require('./protocol/http.js');
 }
 
 // inherit eventEmitter
@@ -4171,7 +4181,7 @@ conn.prototype.run = function(pipedCommands) {
 	global.iobioClients = global.iobioClients || []
 	global.iobioClients.push(this.runner);
 
-	// bind stream events	
+	// bind stream events
 	require('./utils/bindStreamEvents')(this,this.runner);
 }
 
@@ -4190,30 +4200,32 @@ var ws = function(urlBuilder, pipedCommands, opts) {
 	// Call EventEmitter constructor
 	EventEmitter.call(this);
 
-	var wsUrl = 'ws://' + urlBuilder.uri,
+	var protocol = opts.ssl ? 'wss' : 'ws',
+		wsUrl = protocol + '://' + urlBuilder.uri,
 		BinaryClient = require('binaryjs').BinaryClient,
 		client = BinaryClient(wsUrl),
-		me = this;  
+		me = this;
 
 		this.client = client;
-		this.stream;            
+		this.stream;
 
 		client.on('open', function(stream){
-			var stream = client.createStream({event:'run', params : {'url':wsUrl}});    
+			var stream = client.createStream({event:'run', params : {'url':wsUrl}}),
+				first = true;
 			me.stream = stream;
 
 			stream.on('createClientConnection', function(connection) {
-				// determine serverAddress 
-				var serverAddress;				
+				// determine serverAddress
+				var serverAddress;
 				var cmd = pipedCommands[connection.id];
 				if (cmd) {
 					var cmdOpts = cmd.options;
-					var cmdUrlBuilder = cmd.connection.urlBuilder;					
+					var cmdUrlBuilder = cmd.connection.urlBuilder;
 				} else {
 					var cmdOpts =  opts;
-					var cmdUrlBuilder =  urlBuilder;				
+					var cmdUrlBuilder =  urlBuilder;
 				}
-				
+
 
 				// go through by priority
 				if (connection.serverAddress)  // defined by requesting iobio service
@@ -4221,48 +4233,56 @@ var ws = function(urlBuilder, pipedCommands, opts) {
 				else if (cmdOpts && cmdOpts.writeStream && cmdOpts.writeStream.serverAddress) // defined by writestream on client
 					serverAddress = cmdOpts.writeStream.serverAddress
 				else  // defined by client
-					serverAddress = cmdUrlBuilder.getService();				
-				
+					serverAddress = cmdUrlBuilder.getService();
+
 				// connect to server
-				var dataClient = BinaryClient('ws://' + serverAddress);
-				dataClient.on('open', function() {										
+				var dataClient = BinaryClient(protocol + '://' + serverAddress);
+				dataClient.on('open', function() {
 					var dataStream = dataClient.createStream({event:'clientConnected', 'connectionID' : connection.id});
 					var file = cmdUrlBuilder.getFile();
-					file.write(dataStream, cmdOpts);					
+					file.write(dataStream, cmdOpts);
 				})
-            })      
-			
-			stream.on('data', function(data, options) {				
+            })
+
+			stream.on('data', function(data, options) {
+				if(first) { first=false; me.emit('start'); }
 				me.emit('data', data);
 			});
 
 			stream.on('end', function() {
 				me.emit('end');
-			})  
+			})
 
 			stream.on('exit', function(code) {
 				me.emit('exit', code);
-			})   
+			})
 
 			stream.on('error', function(error) {
 				me.emit('error', error);
-			})  
+			})
 
-			stream.on('queue', function(queue) {				
+			stream.on('queue', function(queue) {
 				me.emit('queue', queue)
-			})    
+			})
 		});
 }
 
 // inherit eventEmitter
 inherits(ws, EventEmitter);
 
+ws.prototype.closeClient = function() {
+	if (this.client)
+		this.client.close();
+}
+
 ws.prototype.kill = function() {
-	this.stream.destroy();	
+	if (this.stream)
+		this.stream.destroy();
 }
 
 ws.prototype.end = function() {
-	this.stream.end();	
+	if (this.stream)
+		this.stream.end();
 }
 
 module.exports = ws;
@@ -4350,12 +4370,13 @@ var bindStreamEvents = function(parent, child) {
 	// handle events
 	child.on('data', function(data) { parent.emit('data',data)});
 	child.on('end',   function() { parent.emit('end')});
+	child.on('start', function() { parent.emit('start')});
 	child.on('exit', function(code) { parent.emit('exit',code)});
-	child.on('error', function(error) { parent.emit('error',error)});	
-	child.on('queue', function(queue) { parent.emit('queue', queue)});	
+	child.on('error', function(error) { parent.emit('error',error)});
+	child.on('queue', function(queue) { parent.emit('queue', queue)});
 }
 
-module.exports = bindStreamEvents;	
+module.exports = bindStreamEvents;
 },{}],25:[function(require,module,exports){
 var urlParams = function(params) {
 	var str = ''
